requirements.txt
torch==0.4.1
torchnet==0.0.4
torchvision==0.2.1
easydict==1.9
tqdm==4.29.0
reprint==0.5.2
plyfile==0.7
tensorflow-gpu==1.4.0
pyyaml==3.13
ipdb==0.11
matplotlib==3.0.2
h5py==2.9.0
scikit-image==0.14.2
cffi==1.11.5
